Balls clases changing...

+++ LAB.1/LAB1Ballscollision/app/src/main/java/com/example/lab1ballscollision/objects/Ball.java
@@ -2,9 +2,9 @@ package com.example.lab1ballscollision.objects;
 
 abstract class Ball {
     Vector position, impulse;
-    float mass;
+    double mass;
 
-    Ball(double x, double y, float m) {
+    Ball(double x, double y, double m) {
         position = new Vector(x, y);
         impulse = new Vector();
         mass = m;
